Make tests independent from number of used decimals in the DBF values
Adjust decimals in tests.

More adjustments in decimals

Test error solved?

Perhaps the test would not be done like that. The generated DBF it's not the same than the original one because of the logic of the decimals.

Using round I avoid precision problems. The important thing is to assert the value, not the precision